学习

+++ app/playdatacollect-collector/src/main/java/xie/playdatacollect/collector/quartz/job/DummyJob.java
@@ -9,7 +9,7 @@ public class DummyJob extends XBaseQuartzJobBean {
 		logger.warn("this is dummy job, {}", context.getJobDetail().getKey());
 		System.out.println("this is dummy job, sleep " + context.getMergedJobDataMap().get("sleep"));
 
-		Long time = context.getMergedJobDataMap().containsKey("sleep") ? context.getMergedJobDataMap().getLong("sleep") : 10L;
+		Long time = context.getMergedJobDataMap().containsKey("sleep") ? context.getMergedJobDataMap().getInt("sleep") : 10L;
 		if (time == null) {
 			logger.warn("time is null");
 			time = 10L;
